Extract godoc for datastoredb in a doc.go
